tourism_02/requirements.txt
Flask==3.0.2
firebase-admin==6.4.0
python-dotenv==1.0.1
requests==2.31.0
twilio==8.12.0
Pillow>=10.2.0
gTTS==2.5.3
